Created separate project for message handlers and moved oof reaction handler
