+++ DiscordBot.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DiscordBot.Commands", "Disc
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DiscordBot.Core", "DiscordBot.Core\DiscordBot.Core.csproj", "{BFB8EAF5-4D9A-4610-9AA0-7CA27CE6F201}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordBot.MessageHandlers", "DiscordBot.MessageHandlers\DiscordBot.MessageHandlers.csproj", "{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +53,14 @@ Global
 		{BFB8EAF5-4D9A-4610-9AA0-7CA27CE6F201}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BFB8EAF5-4D9A-4610-9AA0-7CA27CE6F201}.Release|x64.ActiveCfg = Release|x64
 		{BFB8EAF5-4D9A-4610-9AA0-7CA27CE6F201}.Release|x64.Build.0 = Release|x64
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Debug|x64.Build.0 = Debug|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Release|x64.ActiveCfg = Release|Any CPU
+		{A9FF9F8A-B69D-4F85-BA7C-802875A329D2}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
