+++ DiscordBot/Startup.cs
@@ -5,9 +5,9 @@ using DiscordBot.Commands;
 using DiscordBot.Commands.Core;
 using DiscordBot.Core;
 using DiscordBot.Core.Helpers;
-using DiscordBot.Core.MessageReceivedHandlers;
 using DiscordBot.Core.Providers;
 using DiscordBot.Core.Voice;
+using DiscordBot.MessageHandlers;
 using Serilog;
 using Serilog.Extensions.Autofac.DependencyInjection;
 using System.Reflection;
